operations builder in separate class, breaking operations builder in modules
